Rewrite middleware
Added rewrite middleware which accepts a set of rules and transforms the
incoming scoped URL.  This is meant to be used as Pre middleware.
