+++ server_test.go
@@ -169,47 +169,6 @@ func TestNotFound(t *testing.T) {
 	res.Body.Close()
 }
 
-func TestRewrite(t *testing.T) {
-	server := New()
-	server.Route("GET", "/users/:id/profile", func(c Context) Response {
-		return c.R(map[string]interface{}{"id": c.URLParams.Int("id")})
-	})
-
-	server.Rewrite(RewriteRules{
-		"/people/(.*)/profile": "/users/$1/profile",
-	})
-
-	ts := httptest.NewServer(server)
-	defer ts.Close()
-
-	res, err := http.Get(ts.URL + "/people/3/profile")
-	if err != nil {
-		t.Errorf("Error requesting url: %s", err.Error())
-	}
-
-	defer res.Body.Close()
-	bbody, err := ioutil.ReadAll(res.Body)
-	if err != nil {
-		t.Errorf("Error reading response: %s", err.Error())
-	}
-
-	jsRes := struct {
-		Data struct {
-			ID int `json:"id" validate:"eq=3"`
-		}
-	}{}
-
-	err = json.Unmarshal(bbody, &jsRes)
-	if err != nil {
-		t.Error(err.Error())
-		return
-	}
-
-	if err := validator.New().Struct(jsRes); err != nil {
-		t.Error(err.Error())
-	}
-}
-
 func TestDataExtration(t *testing.T) {
 	server := New()
 	server.Route(POST, "/foo", func(c Context) Response {
